feat: animations to themeselector
this is so completely and utterly excessive, this is great. i love adding component libraries to have something accessible.

Also sidenote; the bundler says that melt is taking 60 kb, which is more than React and Radix combined. The package does say it's in beta, but I *hope* tree shaking is actually happening here.