fix: inconsistent article widths on larger screens
On larger screens, the table of contents may change width. Although intentional, this unintentionally caused the whole article section to change widths on larger screens, where it'll always not expand to the width of the navbar. This has been changed to extend to the same width as the navbar when the screen allows for it.

This commit also fixes the draft warning callout's text clipping out of the callout.